OPEN - issue GWDP-375: Alter single-site download to used multi-site pipeline for consistency http://internal.cida.usgs.gov/jira/browse/GWDP-375
Done.
Single-site download will currently download all data types, even if no data is available. This is most like what the Excel download does (with a sheet for each data type, even if not populated).

It would take one-half to one day to omit the CSV files for data type(s) known to not be available.